+++ Api/Controllers/Videos.cs
@@ -1,8 +1,10 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.Extensions.Logging;
 using MySql.Data.MySqlClient;
 using Newtonsoft.Json;
 using Smort_api.Handlers;
+using Smort_api.Object.Video;
 using Smort_api.Object.Videos;
 using System.Security.Claims;
 
@@ -38,7 +40,7 @@ namespace Tiktok_api.Controllers
             if (Data == null)
                 return Task.FromResult("Video Failed to save");
 
-            Data.FileName = $"{videoSavedId}.mkv";
+            Data.FileName = $"{videoSavedId}";
 
             if (Data.MediaData == null)
                 return Task.FromResult("No Video");
@@ -47,19 +49,45 @@ namespace Tiktok_api.Controllers
 
             using DatabaseHandler databaseHandler = new DatabaseHandler();
 
+            int ThumbnailID = 0;
+
+            //Creates File
+            if (Data.Thumbnail != null) { 
+                MySqlCommand GetAndAddProfilePicture = new MySqlCommand();
+                GetAndAddProfilePicture.CommandText =
+                    @"INSERT INTO File (File_Name, File_Location, Created_At) VALUES (@Name, @Location, @Created);
+                      SELECT LAST_INSERT_ID();";
+
+                GetAndAddProfilePicture.Parameters.AddWithValue("@Name", $"{Data.FileName}.png");
+                GetAndAddProfilePicture.Parameters.AddWithValue("@Location", $"./Videos/{id}/{Data.FileName}.png");
+                GetAndAddProfilePicture.Parameters.AddWithValue("@Created", DateTime.Now);
+
+                ThumbnailID = databaseHandler.GetNumber(GetAndAddProfilePicture);
+
+                //SaveThumbnail
+                ImageHandler.SaveThumbnail(Data.Thumbnail, $"{Data.FileName}.png", id.ToString());
+            }
+
+
             using MySqlCommand InsertFileAndVideo = new MySqlCommand();
 
             InsertFileAndVideo.CommandText =
                 @"INSERT INTO File (File_Name, File_location, Created_At, Deleted_At) VALUES (@FileName, @FileLocation, @CreatedAt, @DeletedAt);
-                  INSERT INTO Video (User_Id, File_Id, Title, Description, Created_At, Updated_At, Deleted_At) VALUES (@Id, LAST_INSERT_ID(), @Title, @Description, @CreatedAt, @UpdatedAt, @DeletedAt);";
-           
-            InsertFileAndVideo.Parameters.AddWithValue("@FileName", Data.FileName);
+                  INSERT INTO Video (User_Id, File_Id, Title, Description, Thumbnail, Created_At, Updated_At, Deleted_At) VALUES (@Id, LAST_INSERT_ID(), @Title, @Description, @Thumbnail, @CreatedAt, @UpdatedAt, @DeletedAt);";
+
+            InsertFileAndVideo.Parameters.AddWithValue("@FileName", $"{Data.FileName}.mkv");
             InsertFileAndVideo.Parameters.AddWithValue("@Id", id);
-            InsertFileAndVideo.Parameters.AddWithValue("@FileLocation", $"./Videos/{id}/{Data.FileName}");
+            InsertFileAndVideo.Parameters.AddWithValue("@FileLocation", $"./Videos/{id}/{Data.FileName}.mkv");
             InsertFileAndVideo.Parameters.AddWithValue("@CreatedAt", DateTime.Now);
             InsertFileAndVideo.Parameters.AddWithValue("@DeletedAt", DateTime.Now);
             InsertFileAndVideo.Parameters.AddWithValue("@UpdatedAt", DateTime.Now);
+
             InsertFileAndVideo.Parameters.AddWithValue("@Title", Data.Title);
+
+            Logger.LogInformation(ThumbnailID.ToString());
+
+            InsertFileAndVideo.Parameters.AddWithValue("@Thumbnail", ThumbnailID.ToString());
+
             InsertFileAndVideo.Parameters.AddWithValue("@Description", Data.Description);
 
             using MySqlCommand GetIdVideo = new MySqlCommand();
@@ -122,7 +150,7 @@ namespace Tiktok_api.Controllers
         [HttpGet]
         public ActionResult? GetVideos(int videoId)
         {
-            try 
+            try
             {
                 using MySqlCommand GetVideoPath = new MySqlCommand();
 
@@ -133,7 +161,7 @@ namespace Tiktok_api.Controllers
 
                 string json = databaseHandler.Select(GetVideoPath);
 
-                VideoPathData[] path = JsonConvert.DeserializeObject<VideoPathData[]>(json)!;
+                FilePathData[] path = JsonConvert.DeserializeObject<FilePathData[]>(json)!;
 
                 var filestream = System.IO.File.OpenRead(path[0].File_Location!);
                 return File(filestream, contentType: "video/mkv", enableRangeProcessing: true);
@@ -159,17 +187,16 @@ namespace Tiktok_api.Controllers
             {
                 using MySqlCommand GetVideoPath = new MySqlCommand();
 
-                GetVideoPath.CommandText = "SELECT File_Location FROM File WHERE Id=(SELECT File_Id FROM Video WHERE Id=@Id);";
+                GetVideoPath.CommandText = "SELECT Thumbnail FROM Video WHERE Id=@Id;";
                 GetVideoPath.Parameters.AddWithValue("@Id", $"{videoId}");
 
                 using DatabaseHandler databaseHandler = new DatabaseHandler();
 
                 string json = databaseHandler.Select(GetVideoPath);
 
-                VideoPathData[] path = JsonConvert.DeserializeObject<VideoPathData[]>(json)!;
+                ThumbnailData[] path = JsonConvert.DeserializeObject<ThumbnailData[]>(json)!;
 
-                var filestream = System.IO.File.OpenRead(path[0].File_Location!);
-                return File(filestream, contentType: "Image/png", enableRangeProcessing: true);
+                return File(path[0].Thumbnail, contentType: "Image/png", enableRangeProcessing: false);
 
             }
             catch (Exception ex)
@@ -178,6 +205,7 @@ namespace Tiktok_api.Controllers
                 return BadRequest();
             }
         }
+
 
         /// <summary>
         /// Needs a video Id and removes a video from the database and from the file structure
@@ -212,8 +240,8 @@ namespace Tiktok_api.Controllers
             {
                 string json = databaseHandler.Select(SelectVideoPath);
 
-                VideoPathData[] path = JsonConvert.DeserializeObject<VideoPathData[]>(json)!;
-                if(path.Length != 1)
+                FilePathData[] path = JsonConvert.DeserializeObject<FilePathData[]>(json)!;
+                if (path.Length != 1)
                     return Task.FromResult<ActionResult>(BadRequest());
 
                 System.IO.File.Delete(path[0].File_Location!);
@@ -222,6 +250,85 @@ namespace Tiktok_api.Controllers
             }
 
             return Task.FromResult<ActionResult>(Ok());
+        }
+        /// <summary>
+        /// Gets a list of Thumbnails that a user has made
+        /// </summary>
+        /// <param name="UserId"></param>
+        /// <returns></returns>
+        [Route("Video/GetThumbnailUserList")]
+        [HttpGet]
+        public Task<string> GetThumbnailUserList(int id)
+        {
+
+            try
+            {
+                using MySqlCommand GetVideoPath = new MySqlCommand();
+
+                GetVideoPath.CommandText = "SELECT Id, Title, Thumbnail FROM Video WHERE User_Id=@Id;";
+                GetVideoPath.Parameters.AddWithValue("@Id", $"{id}");
+
+                using DatabaseHandler databaseHandler = new DatabaseHandler();
+
+                return Task.FromResult(databaseHandler.Select(GetVideoPath));
+
+            }
+            catch (Exception ex)
+            {
+                Logger.LogError(ex.Message);
+                return Task.FromResult("");
+            }
+        }
+
+        [Authorize]
+        [Route("Video/GetMyThumbnail")]
+        [HttpGet]
+        public Task<string> GetMyThumbnail()
+        {
+            string token = HttpContext.Request.Headers["Authorization"]!;
+
+            if (JWTTokenHandler.IsBlacklisted(token))
+                return Task.FromResult<string>("Error 405");
+
+            string id = User.FindFirstValue("Id");
+
+            try
+            {
+                using MySqlCommand GetVideoPath = new MySqlCommand();
+
+                GetVideoPath.CommandText = "SELECT Id, Title, Thumbnail FROM Video WHERE User_Id=@Id;";
+                GetVideoPath.Parameters.AddWithValue("@Id", $"{id}");
+
+                using DatabaseHandler databaseHandler = new DatabaseHandler();
+
+                return Task.FromResult(databaseHandler.Select(GetVideoPath));
+
+            }
+            catch (Exception ex)
+            {
+                Logger.LogError(ex.Message);
+                return Task.FromResult("");
+            }
+        }
+
+        /// <summary>
+        /// returns the data of 1 random videos
+        /// </summary>
+        /// <param name="videoId"></param>
+        /// <returns></returns>
+        [Route("Video/GetVideoFromId")]
+        [HttpGet]
+        public Task<string> GetVideoFromId(int id)
+        {
+            using MySqlCommand GetVideo = new MySqlCommand();
+            GetVideo.CommandText = "SELECT Id, Title, Description, Created_At FROM Video WHERE Id=@Id;";
+            GetVideo.Parameters.AddWithValue("@Id", id);
+            using (DatabaseHandler databaseHandler = new DatabaseHandler())
+            {
+                string json = databaseHandler.Select(GetVideo);
+                Logger.LogInformation(json);
+                return Task.FromResult(json);
+            }
         }
     }
 }
