+++ Smort-api.Object/User/GetMyUserDataSimpel.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Smort_api.Object.User
+{
+    public class GetMyUserDataSimpel
+    {
+        public string Username { get; set; }
+        public int Profile_Picture { get; set; }
+    }
+}
