+++ Smort-api.Object/Video/Thumbnail.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Smort_api.Object.Video
+{
+    public class ThumbnailData
+    {
+        public byte[] Thumbnail { get; set; } = new byte[0];
+    }
+}
