+++ Smort-api.Object/Video/VideoUploadData.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using Newtonsoft.Json;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -8,9 +9,14 @@ namespace Smort_api.Object.Videos
 {
     public class VideoUploadData
     {
+        [JsonProperty("key1")]
         public string? FileName { get; set; }
+
+        [JsonProperty("key2")]
         public byte[]? MediaData { get; set; }
         public string? Title { get; set; }
         public string? Description { get; set; }
+
+        public byte[]? Thumbnail { get; set; }
     }
 }
